+++ test/integration/default/default_spec.rb
@@ -30,3 +30,11 @@ describe service('grafana-server') do
   it { should be_enabled }
   it { should be_running }
 end
+
+describe json(content: http('http://localhost:3000/api/admin/settings',
+              auth: { user: 'admin', pass: 'admin' },
+              params: { format: 'html' },
+              method: 'GET',
+              headers: { 'Content-Type' => 'application/json' }).body) do
+  its(['external_image_storage.s3', 'bucket']) { should eq 'grafana-image-store' }
+end