+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This file is used to list changes made in each version of grafana.
 
+## Unreleased
+- After modifying config files, grafana-server is restarted
+
 ## 5.0.0 (2019-06-20)
 
 - Allow custom database names