+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file is used to list changes made in each version of grafana.
 
+## Unreleased
+
+- Fixed session `provider`
+- Added remote-cache config support
+
 ## 5.0.0 (2019-06-20)
 
 - Allow custom database names