+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This file is used to list changes made in each version of grafana.
 ## Unreleased
 
 - After modifying config files, grafana-server is restarted
+- Fixed session `provider`
+- Added remote-cache config support
 
 ## 5.0.0 (2019-06-20)